Tajgram Ultimate: Баробарсозии мукаммал бо DrKLO (Охирин тағйиротҳо дар main)

+++ TMessagesProj/jni/voip/tgcalls/platform/darwin/TGRTCVideoDecoderH264.mm
@@ -67,7 +67,7 @@ static void decompressionOutputCallback(void *decoderRef,
       initWithBuffer:frameBuffer
             rotation:RTCVideoRotation_0
          timeStampNs:CMTimeGetSeconds(timestamp) * rtc::kNumNanosecsPerSec];
-  decodedFrame.timeStamp = (int32_t)decodeParams->timestamp;
+  decodedFrame.timeStamp = decodeParams->timestamp;
   decodeParams->callback(decodedFrame);
   }
 }
